add demos folder
